youtube-summarizer/requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
langdetect==1.0.9
google-api-python-client==2.118.0
textblob==0.17.1 